Massive sweeping changes, and a radically different structuring.
Should be more intuitive to understand how it works as well as to work with.
Still a WIP, as the socket accept and thread handling is not yet complete.
I am attempting to have the scheduler automatically pick up dropped sockets/
spawn the correct number of sockets as needed. Robustness changes, etc.